+++ app/components/Base/RemoteImage/index.js
@@ -1,19 +1,34 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import PropTypes from 'prop-types';
-import { View, Image, ViewPropTypes } from 'react-native';
+import { Image, ViewPropTypes, View } from 'react-native';
 import FadeIn from 'react-native-fade-in-image';
 // eslint-disable-next-line import/default
 import resolveAssetSource from 'react-native/Libraries/Image/resolveAssetSource';
 import { SvgCssUri } from 'react-native-svg';
+import isUrl from 'is-url';
 import ComponentErrorBoundary from '../../UI/ComponentErrorBoundary';
-import useRemoteResourceExists from '../hooks/useRemoteResourceExists';
+import useIpfsGateway from '../../hooks/useIpfsGateway';
+import { util } from '@metamask/controllers';
 
 const RemoteImage = (props) => {
 	// Avoid using this component with animated SVG
 	const source = resolveAssetSource(props.source);
-	const [imageExists] = useRemoteResourceExists(source.uri);
+	const isImageUrl = isUrl(props?.source?.uri);
+	const ipfsGateway = useIpfsGateway();
+	const ipfsHash = useMemo(() => {
+		try {
+			const url = new URL(props.source.uri);
+			if (url.protocol !== 'ipfs:') return false;
+			const contentIdentifier = util.getIpfsUrlContentIdentifier(props.source.uri);
+			return contentIdentifier;
+		} catch {
+			return false;
+		}
+	}, [props.source.uri]);
 
-	if (source && source.uri && source.uri.match('.svg')) {
+	const uri = ipfsHash ? `${ipfsGateway}${ipfsHash}` : source.uri;
+
+	if (source && source.uri && source.uri.match('.svg') && (isImageUrl || ipfsHash)) {
 		const style = props.style || {};
 		if (source.__packager_asset && typeof style !== 'number') {
 			if (!style.width) {
@@ -23,10 +38,11 @@ const RemoteImage = (props) => {
 				style.height = source.height;
 			}
 		}
+
 		return (
 			<ComponentErrorBoundary onError={props.onError} componentLabel="RemoteImage-SVG">
 				<View style={style}>
-					<SvgCssUri {...props} uri={imageExists ? source.uri : undefined} width={'100%'} height={'100%'} />
+					<SvgCssUri {...props} uri={uri} width={'100%'} height={'100%'} fill={'black'} />
 				</View>
 			</ComponentErrorBoundary>
 		);
@@ -35,11 +51,11 @@ const RemoteImage = (props) => {
 	if (props.fadeIn) {
 		return (
 			<FadeIn placeholderStyle={props.placeholderStyle}>
-				<Image {...props} source={{ uri: imageExists ? source.uri : undefined }} />
+				<Image {...props} source={{ uri }} />
 			</FadeIn>
 		);
 	}
-	return <Image {...props} source={{ uri: imageExists ? source.uri : undefined }} />;
+	return <Image {...props} source={{ uri }} />;
 };
 
 RemoteImage.propTypes = {
@@ -63,6 +79,10 @@ RemoteImage.propTypes = {
 	 * Called when there is an error
 	 */
 	onError: PropTypes.func,
+	/**
+	 * This is set if we know that an image is remote
+	 */
+	isUrl: PropTypes.bool,
 };
 
 export default RemoteImage;