+++ app/components/UI/CollectibleContracts/index.js
@@ -1,18 +1,19 @@
-import React, { useCallback } from 'react';
+import React, { useEffect, useCallback } from 'react';
 import PropTypes from 'prop-types';
 import { TouchableOpacity, StyleSheet, View, InteractionManager, Image } from 'react-native';
 import { connect } from 'react-redux';
-import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
-import { baseStyles, colors, fontStyles } from '../../../styles/common';
+import { colors, fontStyles } from '../../../styles/common';
 import { strings } from '../../../../locales/i18n';
+import Engine from '../../../core/Engine';
 import CollectibleContractElement from '../CollectibleContractElement';
 import Analytics from '../../../core/Analytics';
 import { ANALYTICS_EVENT_OPTS } from '../../../util/analytics';
 import { favoritesCollectiblesObjectSelector } from '../../../reducers/collectibles';
+import { removeFavoriteCollectible } from '../../../actions/collectibles';
 import Text from '../../Base/Text';
 import AppConstants from '../../../core/AppConstants';
-import StyledButton from '../StyledButton';
 import { toLowerCaseEquals } from '../../../util/general';
+import { compareTokenIds } from '../../../util/tokens';
 
 const styles = StyleSheet.create({
 	wrapper: {
@@ -32,18 +33,19 @@ const styles = StyleSheet.create({
 		justifyContent: 'center',
 	},
 	addText: {
-		fontSize: 15,
+		fontSize: 14,
 		color: colors.blue,
 		...fontStyles.normal,
 	},
 	footer: {
 		flex: 1,
 		paddingBottom: 30,
+		alignItems: 'center',
+		marginTop: 24,
 	},
 	emptyContainer: {
 		flex: 1,
-		marginBottom: 42,
-		marginHorizontal: 56,
+		marginBottom: 18,
 		justifyContent: 'center',
 		alignItems: 'center',
 	},
@@ -57,8 +59,9 @@ const styles = StyleSheet.create({
 		color: colors.grey200,
 	},
 	emptyText: {
-		color: colors.grey200,
+		color: colors.greyAssetVisibility,
 		marginBottom: 8,
+		fontSize: 14,
 	},
 });
 
@@ -66,7 +69,15 @@ const styles = StyleSheet.create({
  * View that renders a list of CollectibleContract
  * also known as ERC-721 Tokens
  */
-const CollectibleContracts = ({ collectibleContracts, collectibles, navigation, favoriteCollectibles }) => {
+const CollectibleContracts = ({
+	selectedAddress,
+	chainId,
+	navigation,
+	collectibleContracts,
+	collectibles,
+	favoriteCollectibles,
+	removeFavoriteCollectible,
+}) => {
 	const onItemPress = useCallback(
 		(collectible, contractName) => {
 			navigation.navigate('CollectiblesDetails', { collectible, contractName });
@@ -74,23 +85,53 @@ const CollectibleContracts = ({ collectibleContracts, collectibles, navigation, 
 		[navigation]
 	);
 
-	const goToAddCollectible = useCallback(() => {
+	/**
+	 *	Method to check the token id data type of the current collectibles.
+	 *
+	 * @param collectible - Collectible object.
+	 * @returns Boolean indicating if the collectible should be updated.
+	 */
+	const shouldUpdateCollectibleMetadata = (collectible) => typeof collectible.tokenId === 'number';
+
+	/**
+	 * Method to updated collectible and avoid backwards compatibility issues.
+	 * @param address - Collectible address.
+	 * @param tokenId - Collectible token ID.
+	 */
+	const updateCollectibleMetadata = (collectible) => {
+		const { CollectiblesController } = Engine.context;
+		const { address, tokenId } = collectible;
+		CollectiblesController.removeCollectible(address, tokenId);
+		if (String(tokenId).includes('e+')) {
+			removeFavoriteCollectible(selectedAddress, chainId, collectible);
+		} else {
+			CollectiblesController.addCollectible(address, String(tokenId));
+		}
+	};
+
+	useEffect(() => {
+		// TO DO: Move this fix to the controllers layer
+		collectibles.forEach((collectible) => {
+			if (shouldUpdateCollectibleMetadata(collectible)) {
+				updateCollectibleMetadata(collectible);
+			}
+		});
+	});
+
+	const goToAddCollectible = () => {
 		navigation.push('AddAsset', { assetType: 'collectible' });
 		InteractionManager.runAfterInteractions(() => {
 			Analytics.trackEvent(ANALYTICS_EVENT_OPTS.WALLET_ADD_COLLECTIBLES);
 		});
-	}, [navigation]);
+	};
 
-	const renderFooter = useCallback(
-		() => (
-			<View style={styles.footer} key={'collectible-contracts-footer'}>
-				<TouchableOpacity style={styles.add} onPress={goToAddCollectible} testID={'add-collectible-button'}>
-					<Icon name="plus" size={16} color={colors.blue} />
-					<Text style={styles.addText}>{strings('wallet.add_collectibles')}</Text>
-				</TouchableOpacity>
-			</View>
-		),
-		[goToAddCollectible]
+	const renderFooter = () => (
+		<View style={styles.footer} key={'collectible-contracts-footer'}>
+			<Text style={styles.emptyText}>{strings('wallet.no_collectibles')}</Text>
+			<TouchableOpacity style={styles.add} onPress={goToAddCollectible} testID={'add-collectible-button'}>
+				<Text style={styles.addText}>{strings('wallet.add_collectibles')}</Text>
+			</TouchableOpacity>
+		</View>
 	);
 
 	const renderCollectibleContract = useCallback(
@@ -114,7 +155,8 @@ const CollectibleContracts = ({ collectibleContracts, collectibles, navigation, 
 	const renderFavoriteCollectibles = useCallback(() => {
 		const filteredCollectibles = favoriteCollectibles.map((collectible) =>
 			collectibles.find(
-				({ tokenId, address }) => collectible.tokenId === tokenId && collectible.address === address
+				({ tokenId, address }) =>
+					compareTokenIds(collectible.tokenId, tokenId) && collectible.address === address
 			)
 		);
 		return (
@@ -135,10 +177,9 @@ const CollectibleContracts = ({ collectibleContracts, collectibles, navigation, 
 			<View>
 				{renderFavoriteCollectibles()}
 				<View>{collectibleContracts?.map((item, index) => renderCollectibleContract(item, index))}</View>
-				{renderFooter()}
 			</View>
 		),
-		[collectibleContracts, renderFavoriteCollectibles, renderCollectibleContract, renderFooter]
+		[collectibleContracts, renderFavoriteCollectibles, renderCollectibleContract]
 	);
 
 	const goToLearnMore = () =>
@@ -152,21 +193,10 @@ const CollectibleContracts = ({ collectibleContracts, collectibles, navigation, 
 					source={require('../../../images/no-nfts-placeholder.png')}
 					resizeMode={'contain'}
 				/>
-				<Text center style={[styles.emptyTitleText, styles.emptySectionText]}>
-					{strings('wallet.no_nfts_to_show')}
+				<Text center style={styles.emptyTitleText} bold>
+					{strings('wallet.no_nfts_yet')}
 				</Text>
-
-				<Text center style={[styles.emptyText, styles.emptySectionText]}>
-					{strings('wallet.no_collectibles')}
-				</Text>
-				<StyledButton
-					type={'blue'}
-					onPress={goToAddCollectible}
-					containerStyle={[baseStyles.flexGrow, styles.emptySectionText]}
-				>
-					{strings('wallet.manually_import_nfts')}
-				</StyledButton>
-				<Text center big link style={styles.emptySectionText} onPress={goToLearnMore}>
+				<Text center big link onPress={goToLearnMore}>
 					{strings('wallet.learn_more')}
 				</Text>
 			</View>
@@ -176,11 +206,20 @@ const CollectibleContracts = ({ collectibleContracts, collectibles, navigation, 
 	return (
 		<View style={styles.wrapper} testID={'collectible-contracts'}>
 			{collectibles.length ? renderList() : renderEmpty()}
+			{renderFooter()}
 		</View>
 	);
 };
 
 CollectibleContracts.propTypes = {
+	/**
+	 * Chain id
+	 */
+	chainId: PropTypes.string,
+	/**
+	 * Selected address
+	 */
+	selectedAddress: PropTypes.string,
 	/**
 	 * Array of collectibleContract objects
 	 */
@@ -198,12 +237,23 @@ CollectibleContracts.propTypes = {
 	 * Object of collectibles
 	 */
 	favoriteCollectibles: PropTypes.array,
+	/**
+	 * Dispatch remove collectible from favorites action
+	 */
+	removeFavoriteCollectible: PropTypes.func,
 };
 
 const mapStateToProps = (state) => ({
+	chainId: state.engine.backgroundState.NetworkController.provider.chainId,
+	selectedAddress: state.engine.backgroundState.PreferencesController.selectedAddress,
 	collectibleContracts: state.engine.backgroundState.CollectiblesController.collectibleContracts,
 	collectibles: state.engine.backgroundState.CollectiblesController.collectibles,
 	favoriteCollectibles: favoritesCollectiblesObjectSelector(state),
 });
 
-export default connect(mapStateToProps)(CollectibleContracts);
+const mapDispatchToProps = (dispatch) => ({
+	removeFavoriteCollectible: (selectedAddress, chainId, collectible) =>
+		dispatch(removeFavoriteCollectible(selectedAddress, chainId, collectible)),
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(CollectibleContracts);