+++ app/components/Views/ImportFromSeed/index.test.tsx
@@ -8,8 +8,8 @@ const mockStore = configureMockStore();
 const initialState = {
 	user: {
 		passwordSet: true,
-		seedphraseBackedUp: false
-	}
+		seedphraseBackedUp: false,
+	},
 };
 const store = mockStore(initialState);